move iqeEnablement into utils dir
